# Speed up timestamptz decoding