--- a/encode.go
+++ b/encode.go
@@ -51,7 +51,7 @@ func decode(parameterStatus *parameterStatus, s []byte, typ oid.Oid) interface{}
 	case oid.T_bytea:
 		return parseBytea(s)
 	case oid.T_timestamptz:
-		return parseTs(parameterStatus.currentLocation, string(s))
+		return decodeTimestamptzISO(s, parameterStatus.currentLocation)
 	case oid.T_timestamp, oid.T_date:
 		return parseTs(nil, string(s))
 	case oid.T_time:
@@ -205,6 +205,94 @@ func (c *locationCache) getLocation(offset int) *time.Location {
 	}
 
 	return location
+}
+
+// Decode a Time from the "ISO" format.
+func decodeTimestamptzISO(src []byte, sessionLocation *time.Location) time.Time {
+	atoi := func(s []byte) (result int) {
+		for i := 0; i < len(s); i++ {
+			result = result*10 + int(s[i]-'0')
+		}
+		return
+	}
+
+	sepYearMonth := bytes.IndexByte(src, '-')
+	year := atoi(src[:sepYearMonth])
+	src = src[sepYearMonth:]
+
+	// Skips a separator and converts two digits
+	nextTwoDigits := func() (result int) {
+		result = atoi(src[1:3])
+		src = src[3:]
+		return
+	}
+
+	month := nextTwoDigits()
+	day := nextTwoDigits()
+	hour := nextTwoDigits()
+	minute := nextTwoDigits()
+	second := nextTwoDigits()
+	nanosecond, offset := 0, 0
+
+	// Time before current era is suffixed with BC
+	if src[len(src)-1] == 'C' {
+		// Negate the year and add one.
+		// See http://www.postgresql.org/docs/current/static/datetime-input-rules.html
+		year = 1 - year
+
+		// Strip " BC"
+		src = src[:len(src)-3]
+	}
+
+	// Offset from UTC is formatted ±hh[:mm[:ss]]
+	switch {
+	case len(src) > 6 && src[len(src)-6] == ':':
+		offset += atoi(src[len(src)-2:])
+		src = src[:len(src)-3]
+		fallthrough
+
+	case len(src) > 3 && src[len(src)-3] == ':':
+		offset += 60 * atoi(src[len(src)-2:])
+		src = src[:len(src)-3]
+		fallthrough
+
+	default:
+		offset += 3600 * atoi(src[len(src)-2:])
+		if src[len(src)-3] == '-' {
+			offset = -offset
+		}
+		src = src[:len(src)-3]
+	}
+
+	// Fractional seconds
+	if len(src) > 1 {
+		// Skip fraction separator
+		i := 1
+		for ; i < len(src); i++ {
+			nanosecond = nanosecond*10 + int(src[i]-'0')
+		}
+		// Scale to nanosecnds
+		for ; i < 10; i++ {
+			nanosecond *= 10
+		}
+	}
+
+	result := time.Date(
+		year, time.Month(month), day,
+		hour, minute, second, nanosecond,
+		globalLocationCache.getLocation(offset))
+
+	if sessionLocation != nil {
+		// Set the location based on session TimeZone, but only when it reports
+		// the same offset from UTC.
+		sessionTime := result.In(sessionLocation)
+		_, sessionOffset := sessionTime.Zone()
+		if sessionOffset == offset {
+			result = sessionTime
+		}
+	}
+
+	return result
 }
 
 // This is a time function specific to the Postgres default DateStyle