@cbandy 2015-02-11
I've included assertions, as they are certainly necessary. This implementation is twice as fast with half the allocations:

Before this PR:
```
BenchmarkDecodeTimestamptz       1000000              1202 ns/op              64 B/op          2 allocs/op
BenchmarkDecodeTimestamptzMultiThread    1000000              1191 ns/op              64 B/op          2 allocs/op
```

Without assertions:
```
BenchmarkDecodeTimestamptz       3000000               466 ns/op              32 B/op          1 allocs/op
BenchmarkDecodeTimestamptzMultiThread    3000000               484 ns/op              32 B/op          1 allocs/op
```

After this PR:
```
BenchmarkDecodeTimestamptz       3000000               519 ns/op              32 B/op          1 allocs/op
BenchmarkDecodeTimestamptzMultiThread    3000000               527 ns/op              32 B/op          1 allocs/op
```