@uhoh-itsmaciek 2015-02-11
This looks good. It's unfortunate that this would lead to two very similar code paths for parsing time, but maybe that's worth it. My original code certainly wasn't optimized for performance. I'll merge this if there are no objections.

(oh, and I checked Travis and it's the same tip / 9.4 issue as described in #334)