@johto 2015-02-11
How much can we speed up parseTs() if we avoid the second allocation?  Where is it coming from in the first place?  It's not obvious from just looking at the code.

I'm a bit worried about having this much code to maintain just for parsing timestamps.