@cbandy 2015-02-11
> How much can we speed up parseTs()

Since parseTs() is used only for `timestamp` and `date`, code there around offset/location can probably be jettisoned.

> the second allocation? Where is it coming from in the first place?

Perhaps the conversion to string? This version operates on the byte slice. We also called out to `strconv.Atoi`, which uses the generic `ParseUint`.

> having this much code to maintain just for parsing timestamps

In general, or because I added a function instead of touching parseTs? I tried using [time.Parse more heavily](https://gist.github.com/cbandy/77c104ff37d9c8a1ea7a), but it was slower than parseTs.

> two very similar code paths for parsing time

Parsing other time-ish types is being debated in #329. I focused on `timestamptz` since it seemed to be immune to the debate.