@cbandy 2015-02-15
I've added one more commit that I think makes this function easier to read.

```
HEAD is now at 08ad791... Assert expected bytes as they are processed
BenchmarkDecodeTimestamptz       3000000               514 ns/op              32 B/op          1 allocs/op
BenchmarkDecodeTimestamptzMultiThread    3000000               528 ns/op              32 B/op          1 allocs/op

HEAD is now at 013da70... Reduce some duplication when decoding timestamptz
BenchmarkDecodeTimestamptz       3000000               492 ns/op              32 B/op          1 allocs/op
BenchmarkDecodeTimestamptzMultiThread    3000000               521 ns/op              32 B/op          1 allocs/op
```