@johto 2015-02-17
>> having this much code to maintain just for parsing timestamps

> In general, or because I added a function instead of touching parseTs? I tried using time.Parse more heavily, but it was slower than parseTs.

I don't mind not using time.Parse, but I wouldn't want two distinct hand-written routines for parsing time-related data.  Naively thinking, we'd have one function which:

  1. Parses a date in ISO format.  If the data type is "date", the end of string must be nearby as well.
  2. Parses a time.  If the data type is "without time zone", the end of string must be there.
  3. Parses the UTC offset information.  This must be the end of string.

If we can make that function fast while supporting all three type, things ought to be pretty good.