@cbandy 2015-02-17
> I wouldn't want two distinct hand-written routines for parsing time-related data.

I agree. This PR should probably wait for a resolution to #329.